+++ src/actinia_core/rest/ephemeral_processing_with_export.py
@@ -33,10 +33,10 @@ from copy import deepcopy
 from flask_restful_swagger_2 import swagger
 from actinia_core.rest.ephemeral_processing import EphemeralProcessing
 from actinia_core.rest.resource_base import ResourceBase
-from actinia_core.core.common.redis_interface import enqueue_job
-from actinia_core.core.common.process_object import Process
-from actinia_core.core.common.process_chain import ProcessChainModel
-from actinia_core.core.common.exceptions import AsyncProcessTermination
+from actinia_core.core.redis_interface import enqueue_job
+from actinia_core.core.process_object import Process
+from actinia_core.core.process_chain import ProcessChainModel
+from actinia_core.core.exceptions import AsyncProcessTermination
 from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel
