+++ src/actinia_core/rest/location_management.py
@@ -32,8 +32,8 @@ import os
 import shutil
 import pickle
 from flask_restful_swagger_2 import swagger
-from actinia_core.core.common.app import auth
-from actinia_core.core.common.api_logger import log_api_call
+from actinia_core.core.app import auth
+from actinia_core.core.api_logger import log_api_call
 from actinia_core.rest.user_auth import very_admin_role
 from actinia_core.rest.user_auth import check_user_permissions
 from actinia_core.models.response_models import ProcessingResponseModel
@@ -44,8 +44,8 @@ from actinia_core.models.openapi.location_management import ProjectionInfoModel
 from actinia_core.rest.resource_base import ResourceBase
 from actinia_core.rest.persistent_processing import PersistentProcessing
 from actinia_core.rest.mapset_management import PersistentGetProjectionRegionInfo
-from actinia_core.core.common.redis_interface import enqueue_job
-from actinia_core.core.common.exceptions import AsyncProcessError
+from actinia_core.core.redis_interface import enqueue_job
+from actinia_core.core.exceptions import AsyncProcessError
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert, Carmen Tawalika"
