+++ src/actinia_core/rest/persistent_mapset_merger.py
@@ -29,8 +29,8 @@ from flask import jsonify, make_response
 
 from actinia_core.rest.persistent_processing import PersistentProcessing
 from actinia_core.rest.resource_base import ResourceBase
-from actinia_core.core.common.redis_interface import enqueue_job
-from actinia_core.core.common.exceptions import AsyncProcessError, \
+from actinia_core.core.redis_interface import enqueue_job
+from actinia_core.core.exceptions import AsyncProcessError, \
      AsyncProcessTermination
 
 __license__ = "GPLv3"
