+++ src/actinia_core/rest/raster_colors.py
@@ -32,7 +32,7 @@ from flask import jsonify, make_response
 from .ephemeral_processing import EphemeralProcessing
 from .persistent_processing import PersistentProcessing
 from .resource_base import ResourceBase
-from actinia_core.core.common.redis_interface import enqueue_job
+from actinia_core.core.redis_interface import enqueue_job
 import tempfile
 import os
 import atexit
