+++ src/actinia_core/rest/raster_export.py
@@ -32,7 +32,7 @@ import pickle
 from flask_restful_swagger_2 import swagger
 from .resource_base import ResourceBase
 from .ephemeral_processing_with_export import EphemeralProcessingWithExport
-from actinia_core.core.common.redis_interface import enqueue_job
+from actinia_core.core.redis_interface import enqueue_job
 from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel
