+++ src/actinia_core/rest/resource_base.py
@@ -32,13 +32,13 @@ from flask import make_response, jsonify
 from flask import request, g
 from flask.json import loads as json_loads
 from flask_restful_swagger_2 import Resource
-from actinia_core.core.common.app import auth
-from actinia_core.core.common.app import flask_api
-from actinia_core.core.common.config import global_config
-from actinia_core.core.common.api_logger import log_api_call
-from actinia_core.core.common.messages_logger import MessageLogger
-from actinia_core.core.common.resources_logger import ResourceLogger
-from actinia_core.core.common.resource_data_container import ResourceDataContainer
+from actinia_core.core.app import auth
+from actinia_core.core.app import flask_api
+from actinia_core.core.config import global_config
+from actinia_core.core.api_logger import log_api_call
+from actinia_core.core.messages_logger import MessageLogger
+from actinia_core.core.resources_logger import ResourceLogger
+from actinia_core.core.resource_data_container import ResourceDataContainer
 from actinia_core.models.response_models import ProcessingResponseModel
 from actinia_core.models.response_models import create_response_from_model, ApiInfoModel
 from .resource_streamer import RequestStreamerResource
