+++ src/actinia_core/rest/resource_storage_management.py
@@ -34,15 +34,15 @@ import pickle
 from .ephemeral_processing import EphemeralProcessing
 from .persistent_processing import PersistentProcessing
 from .resource_base import ResourceBase
-from actinia_core.core.common.redis_interface import enqueue_job
-from actinia_core.core.common.process_object import Process
-from actinia_core.core.common.exceptions import AsyncProcessError
+from actinia_core.core.redis_interface import enqueue_job
+from actinia_core.core.process_object import Process
+from actinia_core.core.exceptions import AsyncProcessError
 from actinia_core.models.response_models import \
     StorageResponseModel, StorageModel, ProcessingResponseModel, \
     ProcessingErrorResponseModel
 
-from actinia_core.core.common.app import auth
-from actinia_core.core.common.api_logger import log_api_call
+from actinia_core.core.app import auth
+from actinia_core.core.api_logger import log_api_call
 from .user_auth import very_admin_role
 from .user_auth import check_user_permissions
