+++ src/actinia_core/rest/resource_streamer.py
@@ -28,9 +28,9 @@ from flask import jsonify, make_response
 from flask_restful import Resource
 from flask import send_from_directory
 import os
-from actinia_core.core.common.config import global_config
-from actinia_core.core.common.app import auth
-from actinia_core.core.common.api_logger import log_api_call
+from actinia_core.core.config import global_config
+from actinia_core.core.app import auth
+from actinia_core.core.api_logger import log_api_call
 
 
 __license__ = "GPLv3"
