+++ src/actinia_core/rest/strds_renderer.py
@@ -26,7 +26,7 @@ Raster map renderer
 """
 from flask import jsonify, make_response, Response
 from .ephemeral_processing import EphemeralProcessing
-from actinia_core.core.common.redis_interface import enqueue_job
+from actinia_core.core.redis_interface import enqueue_job
 from .renderer_base import RendererBaseResource, EphemeralRendererBase
 import tempfile
 import os
