+++ src/actinia_core/rest/user_api_key.py
@@ -34,8 +34,8 @@ from flask_restful import reqparse
 from flask_restful_swagger_2 import swagger
 from flask import jsonify, make_response, g
 from .base_login import LoginBase
-from actinia_core.core.common.app import auth
-from actinia_core.core.common.api_logger import log_api_call
+from actinia_core.core.app import auth
+from actinia_core.core.api_logger import log_api_call
 from actinia_core.models.openapi.user_api_key import TokenResponseModel
 
 __license__ = "GPLv3"
